Backend/requirements.txt
fastapi[all]
uvicorn
sqlmodel
motor
asyncpg
python-dotenv
fastapi-users
passlib
python-jose
email-validator
pydantic
pydantic-settings
psycopg2-binary
passlib[bcrypt]

# Dev & Testing
black
ruff
pytest
pytest-asyncio
httpx
